feat: complete Modules(src/modules) task
